+++ src/redturtle/importer/base/transmogrifier/blueprints/source.py
@@ -274,7 +274,8 @@ class CachedCatalogSourceSection(object):
             if json_data:
                 try:
                     item_mod_date = datetime.strptime(
-                        item.get("modification_date")[:-6], "%Y/%m/%d %H:%M:%S.%f",
+                        item.get("modification_date")[:-6],
+                        "%Y/%m/%d %H:%M:%S.%f",
                     )
                 except ValueError:
                     #  date in different format
@@ -283,12 +284,14 @@ class CachedCatalogSourceSection(object):
                     )
                 try:
                     item_cache_mod_date = datetime.strptime(
-                        json_data.get("modification_date")[:-6], "%Y/%m/%d %H:%M:%S.%f",
+                        json_data.get("modification_date")[:-6],
+                        "%Y/%m/%d %H:%M:%S.%f",
                     )
                 except ValueError:
                     # date in different format
                     item_cache_mod_date = datetime.strptime(
-                        json_data.get("modification_date")[:-6], "%Y/%m/%d %H:%M:%S",
+                        json_data.get("modification_date")[:-6],
+                        "%Y/%m/%d %H:%M:%S",
                     )
                 if item_mod_date <= item_cache_mod_date:
                     logger.info("HIT path: {0}".format(path))