+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 
 - Better handle datetime and date values.
   [cekk]
-
+- Make transmogrifier more configurable.
+  [cekk]
 
 2.0.2 (2021-02-01)
 ------------------